refactor: improve code quality and robustness
- Fix rate limiting memory issue: use while loop with pop_front() instead of single erase
- Fix sequence_number type mismatch: use int64_t consistently between protocol and session
- Add thread safety to logger fallback path with dedicated mutex
- Add noexcept specifiers to connection_manager default constructor/destructor
- Add explicit copy/move deletion to websocket_server class
- Improve error messages in extract_message_type with specific error details

+++ src/common/protocol.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <functional>
 #include <optional>
 #include <string>
@@ -75,7 +76,7 @@ struct action_message {
   std::string session_id;
   std::string action_type;  // "FOLD", "CALL", "RAISE", "CHECK", "ALL_IN"
   std::optional<double> amount;
-  int sequence_number;
+  int64_t sequence_number;
 };
 
 // STATE_UPDATE message - Server broadcasts game state
